+++ Changelog.md
@@ -25,6 +25,7 @@
   - Editing `Color?` and `Enum?` values works now
   - Improved resource lookup (used to get resource keys from resources)
   - Added dark theme
+  - [#278](../../issues/278) - Adorner Layer not visible on certain controls (Snoop now reports a diagnostic error when there is no adorner layer for the selected element)
   - [#283](../../issues/283) - [Feature Request] Be able to import filters or make them available across applications. (solved by [#316](../../issues/316))
   - [#314](../../issues/314) - Hide properties from Snoop?
   - [#320](../../issues/320) - System resources are not shown in the tree